backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
aiohttp==3.9.1
aiofiles==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.92.1
python-dotenv==1.0.0
pyyaml==6.0.1
